feat: move rate limiter types into own files
